% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{mrmoss_check_formatted_inputs}
\alias{mrmoss_check_formatted_inputs}
\title{Check Formatted GWAS Input Files}
\usage{
mrmoss_check_formatted_inputs(
  formatted_dir,
  traits,
  required_columns = c("SNP", "A1", "A2", "Z", "N", "P")
)
}
\arguments{
\item{formatted_dir}{Directory containing formatted trait files.}

\item{traits}{Character vector of trait names.}

\item{required_columns}{Required column names.}
}
\value{
A data.table summarizing input-file readiness.
}
\description{
Check Formatted GWAS Input Files
}
