% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_clump.R
\name{mrmoss_clump_data}
\alias{mrmoss_clump_data}
\title{LD Clumping for Instrument Selection}
\usage{
mrmoss_clump_data(
  dat,
  iv_threshold,
  reference_prefix = NULL,
  plink_bin = NULL,
  pop = "EUR",
  clump_kb = 1000,
  clump_r2 = 0.001,
  clump_p = 0.999,
  f_min = 10
)
}
\arguments{
\item{dat}{Harmonized data from `mrmoss_harmonize_data`.}

\item{iv_threshold}{IV p-value threshold on exposure.}

\item{reference_prefix}{Prefix to PLINK reference panel (e.g. "/path/EUR").
If \code{NULL}, use the online OpenGWAS clumping service via \code{ieugwasr}.}

\item{plink_bin}{Path to PLINK binary for local clumping.}

\item{pop}{Population code for ieugwasr (default "EUR").}

\item{clump_kb}{Clumping window in kb.}

\item{clump_r2}{Clumping r2 threshold.}

\item{clump_p}{Secondary clumping p threshold.}

\item{f_min}{Minimum F statistic.}
}
\value{
Clumped MR dataset.
}
\description{
LD Clumping for Instrument Selection
}
