% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_clump.R
\name{mrmoss_estimate_outcome_correlation}
\alias{mrmoss_estimate_outcome_correlation}
\title{Estimate Outcome Correlation Matrix from Null SNPs}
\usage{
mrmoss_estimate_outcome_correlation(
  outcome_data_list,
  p_null_threshold = 1e-05
)
}
\arguments{
\item{outcome_data_list}{Named list of formatted outcome data.tables.}

\item{p_null_threshold}{Null-SNP p-value threshold.}
}
\value{
Positive-definite correlation matrix.
}
\description{
Estimate Outcome Correlation Matrix from Null SNPs
}
