% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_clump.R
\name{mrmoss_harmonize_data}
\alias{mrmoss_harmonize_data}
\title{Harmonize Exposure and Outcome Summary Statistics}
\usage{
mrmoss_harmonize_data(dat1, dat2)
}
\arguments{
\item{dat1}{Exposure data with columns SNP/A1/A2/Z/N/P.}

\item{dat2}{Outcome data with columns SNP/A1/A2/Z/N/P.}
}
\value{
Harmonized data for MR estimation.
}
\description{
Harmonize Exposure and Outcome Summary Statistics
}
