% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{mrmoss_load_formatted_traits}
\alias{mrmoss_load_formatted_traits}
\title{Load Formatted Trait Files}
\usage{
mrmoss_load_formatted_traits(formatted_dir, traits)
}
\arguments{
\item{formatted_dir}{Directory containing formatted trait files.}

\item{traits}{Character vector of trait names.}
}
\value{
Named list of data.tables.
}
\description{
Load Formatted Trait Files
}
