% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{mrmoss_read_manifest}
\alias{mrmoss_read_manifest}
\title{Read Manifest for Raw-to-Formatted Conversion}
\usage{
mrmoss_read_manifest(manifest_csv)
}
\arguments{
\item{manifest_csv}{Path to CSV manifest.}
}
\value{
A data.table manifest.
}
\description{
Read Manifest for Raw-to-Formatted Conversion
}
